requirements.txt
google-cloud-bigquery>=3.0
python-dotenv>=1.0
requests>=2.31
